# Implement buffer recycling for CharacterReader