--- a/CHANGES
+++ b/CHANGES
@@ -16,6 +16,10 @@ Release 1.17.1 [PENDING]
     elements to be returned when used on elements other than the root document.
     <https://github.com/jhy/jsoup/issues/2018>
 
+  * Bugfix: in a sub-query such as `p:has(> span, > i)`, combinators following the `,` Or combinator would be
+    incorrectly skipped, such that the sub-query was parsed as `i` instead of `> i`.
+    <https://github.com/jhy/jsoup/issues/1707>
+
 Release 1.16.2 [20-Oct-2023]
   * Improvement: optimized the performance of complex CSS selectors, by adding a cost-based query planner. Evaluators
     are sorted by their relative execution cost, and executed in order of lower to higher cost. This speeds the