--- a/CHANGES
+++ b/CHANGES
@@ -8,6 +8,9 @@ Release 1.17.1 [PENDING]
 
   * Improvement: when changing the OutputSettings syntax to XML, the xhtml EscapeMode is automatically set by default.
 
+  * Improvement: added the `:is(selector list)` pseudo-selector, which finds elements that match any of the selectors in
+    the selector list. Useful for making large ORed selectors more readable.
+
   * Bugfix: when outputting with XML syntax, HTML elements that were parsed as data nodes (<script> and <style>) should
     be emitted as CDATA nodes, so that they can be parsed correctly by an XML parser.
     <https://github.com/jhy/jsoup/pull/1720>