… pom.xml
… src/main/java/org/jsoup/select/QueryParser.java
… src/main/java/org/jsoup/select/Selector.java
… src/main/java/org/jsoup/select/StructuralEvaluator.java
… src/test/java/org/jsoup/benchmarks/JMHBenchmark.java
… src/test/java/org/jsoup/parser/CharacterReaderTest.java
… src/test/java/org/jsoup/select/QueryParserTest.java
… src/test/java/org/jsoup/select/SelectorTest.java