# Before

```
Benchmark                                                  Mode  Cnt      Score      Error   Units
JMHBenchmark.benchmarkSmall                                avgt    5      7.382 ±    0.940   us/op
JMHBenchmark.benchmarkSmall:·gc.alloc.rate                 avgt    5   8572.806 ± 1082.658  MB/sec
JMHBenchmark.benchmarkSmall:·gc.alloc.rate.norm            avgt    5  72952.001 ±    0.001    B/op
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Eden_Space        avgt    5   8776.474 ± 1059.021  MB/sec
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Eden_Space.norm   avgt    5  74688.242 ±  767.418    B/op
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Old_Gen           avgt    5      0.195 ±    0.016  MB/sec
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Old_Gen.norm      avgt    5      1.664 ±    0.174    B/op
JMHBenchmark.benchmarkSmall:·gc.count                      avgt    5    523.000             counts
JMHBenchmark.benchmarkSmall:·gc.time                       avgt    5    761.000                 ms
JMHBenchmark.benchmarkMedium                               avgt    5     29.383 ±    1.479   us/op
JMHBenchmark.benchmarkMedium:·gc.alloc.rate                avgt    5   2531.961 ±  127.589  MB/sec
JMHBenchmark.benchmarkMedium:·gc.alloc.rate.norm           avgt    5  85824.002 ±    0.001    B/op
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Eden_Space       avgt    5   2522.477 ±  106.173  MB/sec
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Eden_Space.norm  avgt    5  85506.478 ± 2628.227    B/op
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Old_Gen          avgt    5      0.207 ±    0.036  MB/sec
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Old_Gen.norm     avgt    5      7.031 ±    1.401    B/op
JMHBenchmark.benchmarkMedium:·gc.count                     avgt    5    372.000             counts
JMHBenchmark.benchmarkMedium:·gc.time                      avgt    5    371.000                 ms
```

# After

```
Benchmark                                                  Mode  Cnt      Score     Error   Units
JMHBenchmark.benchmarkSmall                                avgt    5      2.038 ±   0.102   us/op
JMHBenchmark.benchmarkSmall:·gc.alloc.rate                 avgt    5   3205.214 ± 160.609  MB/sec
JMHBenchmark.benchmarkSmall:·gc.alloc.rate.norm            avgt    5   7536.000 ±   0.001    B/op
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Eden_Space        avgt    5   3175.262 ± 132.074  MB/sec
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Eden_Space.norm   avgt    5   7465.807 ± 139.856    B/op
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Old_Gen           avgt    5      0.194 ±   0.016  MB/sec
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Old_Gen.norm      avgt    5      0.455 ±   0.038    B/op
JMHBenchmark.benchmarkSmall:·gc.count                      avgt    5    439.000            counts
JMHBenchmark.benchmarkSmall:·gc.time                       avgt    5    397.000                ms
JMHBenchmark.benchmarkMedium                               avgt    5     21.085 ±   1.112   us/op
JMHBenchmark.benchmarkMedium:·gc.alloc.rate                avgt    5    833.438 ±  44.777  MB/sec
JMHBenchmark.benchmarkMedium:·gc.alloc.rate.norm           avgt    5  20272.002 ±   0.001    B/op
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Eden_Space       avgt    5    827.537 ±  70.651  MB/sec
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Eden_Space.norm  avgt    5  20126.922 ± 783.745    B/op
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Old_Gen          avgt    5      0.206 ±   0.016  MB/sec
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Old_Gen.norm     avgt    5      5.021 ±   0.162    B/op
JMHBenchmark.benchmarkMedium:·gc.count                     avgt    5    268.000            counts
JMHBenchmark.benchmarkMedium:·gc.time                      avgt    5    193.000                ms
```

# Conclusion

This changeset significantly improves the memory efficiency of JSOUP which turns into massive performance gains.