@jhy 2022-07-01
Good stuff, thanks! I may move some things around (not sure if the recycler should be a public contract API, etc).

Enjoy your vacation :)