@chibenwa 2022-07-01
The next big thing is improving the efficiency of the InputStream based APIs

```
    @Benchmark
    public void benchmarkMediumInputStream(Blackhole bh) throws Exception{
        bh.consume(Jsoup.parse(new ByteArrayInputStream(CONTENT_MEDIUM_AS_BYTES), StandardCharsets.UTF_8.name(), ""));
    }
```

Yields

```
Benchmark                                                             Mode  Cnt       Score      Error   Units
JMHBenchmark.benchmarkMediumInputStream                               avgt    5      43.750 ±   20.967   us/op
JMHBenchmark.benchmarkMediumInputStream:·gc.alloc.rate                avgt    5    2757.207 ± 1220.923  MB/sec
JMHBenchmark.benchmarkMediumInputStream:·gc.alloc.rate.norm           avgt    5  137576.004 ±    0.002    B/op
JMHBenchmark.benchmarkMediumInputStream:·gc.churn.G1_Eden_Space       avgt    5    2761.204 ± 1250.661  MB/sec
JMHBenchmark.benchmarkMediumInputStream:·gc.churn.G1_Eden_Space.norm  avgt    5  137743.285 ± 1963.397    B/op
JMHBenchmark.benchmarkMediumInputStream:·gc.churn.G1_Old_Gen          avgt    5       0.307 ±    0.219  MB/sec
JMHBenchmark.benchmarkMediumInputStream:·gc.churn.G1_Old_Gen.norm     avgt    5      15.431 ±   11.324    B/op
JMHBenchmark.benchmarkMediumInputStream:·gc.count                     avgt    5     326.000             counts
JMHBenchmark.benchmarkMediumInputStream:·gc.time                      avgt    5     489.000                 ms
```

Which memory wise is terrible compared to the strign version... I'd be better of creating the String prior parsing rather than supplying the imputStream....