@chibenwa 2022-07-01
```
Benchmark                                                             Mode  Cnt       Score      Error   Units
JMHBenchmark.benchmarkMedium                                          avgt    5      20.314 ±    1.621   us/op
JMHBenchmark.benchmarkMedium:·gc.alloc.rate                           avgt    5     654.901 ±   51.181  MB/sec
JMHBenchmark.benchmarkMedium:·gc.alloc.rate.norm                      avgt    5   15344.002 ±    0.001    B/op
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Eden_Space                  avgt    5     650.845 ±   57.968  MB/sec
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Eden_Space.norm             avgt    5   15248.663 ±  495.411    B/op
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Old_Gen                     avgt    5       0.193 ±    0.031  MB/sec
JMHBenchmark.benchmarkMedium:·gc.churn.G1_Old_Gen.norm                avgt    5       4.533 ±    0.927    B/op
JMHBenchmark.benchmarkMedium:·gc.count                                avgt    5     244.000             counts
JMHBenchmark.benchmarkMedium:·gc.time                                 avgt    5     174.000                 ms
JMHBenchmark.benchmarkMediumInputStream                               avgt    5      32.855 ±    1.372   us/op
JMHBenchmark.benchmarkMediumInputStream:·gc.alloc.rate                avgt    5    3633.766 ±  150.956  MB/sec
JMHBenchmark.benchmarkMediumInputStream:·gc.alloc.rate.norm           avgt    5  137728.003 ±    0.001    B/op
JMHBenchmark.benchmarkMediumInputStream:·gc.churn.G1_Eden_Space       avgt    5    3632.474 ±  189.650  MB/sec
JMHBenchmark.benchmarkMediumInputStream:·gc.churn.G1_Eden_Space.norm  avgt    5  137676.632 ± 2596.732    B/op
JMHBenchmark.benchmarkMediumInputStream:·gc.churn.G1_Old_Gen          avgt    5       0.339 ±    0.368  MB/sec
JMHBenchmark.benchmarkMediumInputStream:·gc.churn.G1_Old_Gen.norm     avgt    5      12.879 ±   14.070    B/op
JMHBenchmark.benchmarkMediumInputStream:·gc.count                     avgt    5     420.000             counts
JMHBenchmark.benchmarkMediumInputStream:·gc.time                      avgt    5     427.000                 ms
JMHBenchmark.benchmarkSmall                                           avgt    5       1.656 ±    0.035   us/op
JMHBenchmark.benchmarkSmall:·gc.alloc.rate                            avgt    5    1699.708 ±   36.050  MB/sec
JMHBenchmark.benchmarkSmall:·gc.alloc.rate.norm                       avgt    5    3248.000 ±    0.001    B/op
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Eden_Space                   avgt    5    1681.110 ±   63.848  MB/sec
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Eden_Space.norm              avgt    5    3212.422 ±   74.478    B/op
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Old_Gen                      avgt    5       0.186 ±    0.015  MB/sec
JMHBenchmark.benchmarkSmall:·gc.churn.G1_Old_Gen.norm                 avgt    5       0.355 ±    0.034    B/op
JMHBenchmark.benchmarkSmall:·gc.count                                 avgt    5     337.000             counts
JMHBenchmark.benchmarkSmall:·gc.time                                  avgt    5     270.000                 ms
```