@chibenwa 2022-09-13
Hello @jhy , any status on this work?

Also would it make sense to add a JVM system property to turn off the pooling behavior if not desired by the end user?