@jhy 2023-01-06
My apologies for the radio silience @chibenwa, I was pulled away from this.

I have a refactoring of this to make the recyclers more generic, and will push an update for your review shortly.